api/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.0
neo4j==5.25.0
llama-index-llms-gemini==0.3.8
llama-index-embeddings-gemini==0.2.4
llama-index-llms-groq==0.2.6
llama-index-core==0.12.3
rapidfuzz==3.10.0
pydantic==2.10.3
